requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-genai>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pandas>=2.0.0
